+++ python/cdp/openapi_client/api/onramp_api.py
@@ -386,7 +386,6 @@ class OnrampApi:
             '401': "Error",
             '404': "Error",
             '429': "Error",
-            '500': "Error",
         }
         response_data = await self.api_client.call_api(
             *_param,
@@ -457,7 +456,6 @@ class OnrampApi:
             '401': "Error",
             '404': "Error",
             '429': "Error",
-            '500': "Error",
         }
         response_data = await self.api_client.call_api(
             *_param,
@@ -528,7 +526,6 @@ class OnrampApi:
             '401': "Error",
             '404': "Error",
             '429': "Error",
-            '500': "Error",
         }
         response_data = await self.api_client.call_api(
             *_param,